Record finished survey (#3)
* Record survey completion for a single user

* Add index on `SurveyTaker.isFinished`

Apparently, an index is required to be able to use `count` with a
specific field.

* Update survey's state when last survey taker is finished

* Add section of completed surveys on index

* Send updates via websocket when survey response recorded

Add `socket.io` for websockets. On the survey results page, sign up for
updates using the user's id (or token?) and get messages on that
channel. For now, the id is just hardcoded to "someId".

* Convert survey show from Pug to Vue

* Survey updates when there are updates

* Use a token to subscribe to a survey